Changed media-controller a bit to throw error first, all tests working now
