+++ test/server/media-controller.spec.js
@@ -1,8 +1,6 @@
-const chai = require( 'chai' );
-const expect = chai.expect;
-const config = require( '../../app/models/config-model' ).server;
-
-const IPfiltering = config[ 'ip filtering' ];
+const app = require( '../../config/express' );
+const request = require( 'supertest' );
+const http = require( 'http' );
 
 /**
  * Tests the request-filtering-agent to block SSRF attacks
@@ -12,161 +10,246 @@ const IPfiltering = config[ 'ip filtering' ];
  */
 
 const testHTMLBody = 'im in.';
-const enketoHost = 'http://localhost:8005';
-const testHTMLHost = 'http/localhost:1234';
+const portHTML = 1234;
+const testHTMLHost = `http/localhost:${portHTML}`;
+const testHTMLMetaHost = `http/0.0.0.0:${portHTML}`;
+const localhost = '127.0.0.1';
 
-const requestURL = enketoHost + '/media/get/' + testHTMLHost;
+const requestURL = `/media/get/${testHTMLHost}`;
+const requestMetaURL = `/media/get/${testHTMLMetaHost}`;
+const server = http.createServer( function( req, res ) {
+    res.writeHead( 200, { 'Content-Type': 'text/plain' } );
+    res.end( testHTMLBody );
+} );
 
-var http = require('http');
+describe( 'Testing request-filtering-agent', function() {
 
-var server = http.createServer(function (req, res) {
-    res.writeHead(200, { 'Content-Type': 'text/plain' });
-    res.end('im in.');
-});
+    // Default everything disabled
+    const allowPrivateIPAddress = false;
+    const allowMetaIPAddress = false;
+    const allowIPAddressList = [];
+    const denyIPAddressList = [];
 
-describe('Testing request-filtering-agent on the following options:', function () {
+    before( function() {
+        server.listen( portHTML );
+    } );
 
-    before(function () {
-        server.listen(1234);
-    });
-    
-    after(function () {
+    after( function() {
         server.close();
-    });
+    } );
 
-    describe('\n\tallowPrivateIPAddress = ' + IPfiltering.allowPrivateIPAddress 
-        + '\n\tallowMetaIPAddress = ' + IPfiltering.allowMetaIPAddress 
-        + '\n\tallowIPAddressList = ' + IPfiltering.allowIPAddressList 
-        + '\n\tdenyIPAddressList = ' + IPfiltering.denyIPAddressList + '\n' 
-    , function () {
-        it('WITH a Referer', function (done) {
-            const options = {
-                host: '127.0.0.1',
-                port: 8005,
-                path: '/media/get/' + testHTMLHost,
-                headers: {
-                    'Referer': 'https://google.com?print=true'
-                }
-            };
-            if (!IPfiltering.allowPrivateIPAddress) {
-                if (IPfiltering.allowIPAddressList.length == 0 && IPfiltering.denyIPAddressList.length == 0) {
-                    http.get(options, function(){}).on('error', function(err){
-                        expect(err.code).contains('ECONN');
-                    });
-                    done();
-                }
-                else if (IPfiltering.allowIPAddressList.includes( 'localhost' ) || IPfiltering.allowIPAddressList.includes( '127.0.0.1' )) {
-                    http.get(options, function(res) {
-                        var data = '';
-                        res.on('data', function(chunk) {
-                            data+=chunk;
-                        });
-                        res.on('end', function() {
-                            expect(data).to.be.equal(testHTMLBody);
-                            done();
-                        });
-                    });
-                }
-                else if (IPfiltering.denyIPAddressList.includes( 'localhost' ) || IPfiltering.denyIPAddressList.includes( '127.0.0.1' )) {
-                    http.get(options, function(){}).on('error', function(err){
-                        expect(err.code).contains('ECONN');
-                    });
-                    done();
-                }
-            }
-            if (IPfiltering.allowPrivateIPAddress) {
-                if (IPfiltering.allowIPAddressList.length == 0 && IPfiltering.denyIPAddressList.length == 0) {
-                    http.get(options, function(res) {
-                        var data = '';
-                        res.on('data', function(chunk) {
-                            data+=chunk;
-                        });
-                        res.on('end', function() {
-                            expect(data).to.be.equal(testHTMLBody);
-                            done();
-                        });
-                    });
-                }
-                else if (IPfiltering.allowIPAddressList.includes( 'localhost' ) || IPfiltering.allowIPAddressList.includes( '127.0.0.1' )) {
-                    http.get(options, function(res) {
-                        var data = '';
-                        res.on('data', function(chunk) {
-                            data+=chunk;
-                        });
-                        res.on('end', function() {
-                            expect(data).to.be.equal(testHTMLBody);
-                            done();
-                        });
-                    });
-                }
-                else if (IPfiltering.denyIPAddressList.includes( 'localhost' ) || IPfiltering.denyIPAddressList.includes( '127.0.0.1' )) {
-                    http.get(options, function(){}).on('error', function(err){
-                        expect(err.code).contains('ECONN');
-                    });
-                    done();
-                }
-            }
-        });
+    // Tests WITH Referers
 
-        it('WITHOUT a Referer', function (done) {
-            if (!IPfiltering.allowPrivateIPAddress) {
-                if (IPfiltering.allowIPAddressList.length == 0 && IPfiltering.denyIPAddressList.length == 0) {
-                    http.get(requestURL, function(){}).on('error', function(err){
-                        expect(err.code).contains('ECONN');
-                    });
-                    done();
-                }
-                else if (IPfiltering.allowIPAddressList.includes( 'localhost' ) || IPfiltering.allowIPAddressList.includes( '127.0.0.1' )) {
-                    http.get(requestURL, function(res) {
-                        var data = '';
-                        res.on('data', function(chunk) {
-                            data+=chunk;
-                        });
-                        res.on('end', function() {
-                            expect(data).to.be.equal(testHTMLBody);
-                        });
-                        done();
-                    });
-                }
-                else if (IPfiltering.denyIPAddressList.includes( 'localhost' ) || IPfiltering.denyIPAddressList.includes( '127.0.0.1' )) {
-                    http.get(requestURL, function(){}).on('error', function(err){
-                        expect(err.code).contains('ECONN');
-                    });
-                    done();
-                }
-            }
-            if (IPfiltering.allowPrivateIPAddress) {
-                if (IPfiltering.allowIPAddressList.length == 0 && IPfiltering.denyIPAddressList.length == 0) {
-                    http.get(requestURL, function(res) {
-                        var data = '';
-                        res.on('data', function(chunk) {
-                            data+=chunk;
-                        });
-                        res.on('end', function() {
-                            expect(data).to.be.equal(testHTMLBody);
-                        });
-                        done();
-                    });
-                }
-                else if (IPfiltering.allowIPAddressList.includes( 'localhost' ) || IPfiltering.allowIPAddressList.includes( '127.0.0.1' )) {
-                    http.get(requestURL, function(res) {
-                        var data = '';
-                        res.on('data', function(chunk) {
-                            data+=chunk;
-                        });
-                        res.on('end', function() {
-                            expect(data).to.be.equal(testHTMLBody);
-                        });
-                        done();
-                    });
-                }
-                else if (IPfiltering.denyIPAddressList.includes( 'localhost' ) || IPfiltering.denyIPAddressList.includes( '127.0.0.1' )) {
-                    http.get(requestURL, function(){}).on('error', function(err){
-                        expect(err.code).contains('ECONN');
-                    });
-                    done();
-                }
-            }
-        });
-    });
-});
+    // Tests with allowPrivateIPAddress FALSE
+    it( 'for a private IP address WITH a Referer with allowPrivateIPAddress=false', done => {
+
+        // Don't change any default IP filtering setting
+        app.set( 'ip filtering', { allowPrivateIPAddress, allowMetaIPAddress, allowIPAddressList, denyIPAddressList } );
+
+        request( app )
+            .get( requestURL )
+            .set( 'Referer', 'https://google.com?print=true' )
+            .expect( 500, /DNS lookup .* is not allowed. Because, It is private IP address/)
+            .end( done );
+    } );
+    it( 'for a private IP address WITH a Referer with allowPrivateIPAddress=false and allowMetaIPAddress=true', done => {
+ 
+        // Only change one setting
+        const allowMetaIPAddress = true;
+
+        app.set( 'ip filtering', { allowPrivateIPAddress, allowMetaIPAddress, allowIPAddressList, denyIPAddressList } );
+
+        request( app )
+            .get( requestMetaURL  )
+            .set( 'Referer', 'https://google.com?print=true' )
+            .expect( 500, /Socket is closed/ )
+            .end( done );
+    } );
+    it( 'for a private IP address WITH a Referer with allowPrivateIPAddress=false but allowIPAddressList=[`127.0.0.1`]', done => {
+ 
+        // Only change one setting
+        const allowIPAddressList = [localhost];
+
+        app.set( 'ip filtering', { allowPrivateIPAddress, allowMetaIPAddress, allowIPAddressList, denyIPAddressList } );
+
+        request( app )
+            .get( requestURL )
+            .set( 'Referer', 'https://google.com?print=true' )
+            .expect( 200, testHTMLBody )
+            .end( done );
+    } );
+    it( 'for a private IP address WITH a Referer with allowPrivateIPAddress=false and denyIPAddressList=[`127.0.0.1`]', done => {
+
+        // Only change one setting 
+        const denyIPAddressList = [localhost];
+
+        app.set( 'ip filtering', { allowPrivateIPAddress, allowMetaIPAddress, allowIPAddressList, denyIPAddressList } );
+
+        request( app )
+            .get( requestURL )
+            .set( 'Referer', 'https://google.com?print=true' )
+            .expect( 500, /DNS lookup .* is not allowed. Because, It is private IP address/ )
+            .end( done );
+    } );
+
+    // Tests with allowPrivateIPAddress TRUE
+    it( 'for a private IP address WITH a Referer with allowPrivateIPAddress=true', done => {
+
+        // Only change one setting
+        const allowPrivateIPAddress = true;
+
+        app.set( 'ip filtering', { allowPrivateIPAddress, allowMetaIPAddress, allowIPAddressList, denyIPAddressList } );
+
+        request( app )
+            .get( requestURL )
+            .set( 'Referer', 'https://google.com?print=true' )
+            .expect( 200, testHTMLBody )
+            .end( done );
+    } );
+    it( 'for a private IP address WITH a Referer with allowPrivateIPAddress=true and allowMetaIPAddress=true', done => {
+
+        // Change two settings
+        const allowPrivateIPAddress = true;
+        const allowMetaIPAddress = true;
+
+        app.set( 'ip filtering', { allowPrivateIPAddress, allowMetaIPAddress, allowIPAddressList, denyIPAddressList } );
+
+        request( app )
+            .get( requestMetaURL )
+            .set( 'Referer', 'https://google.com?print=true' )
+            .expect( 200, testHTMLBody )
+            .end( done );
+    } );
+    it( 'for a private IP address WITH a Referer with allowPrivateIPAddress=true and allowIPAddressList=[`127.0.0.1`]', done => {
+
+        // Change two settings
+        const allowPrivateIPAddress = true;
+        const allowIPAddressList = [localhost];
+
+        app.set( 'ip filtering', { allowPrivateIPAddress, allowMetaIPAddress, allowIPAddressList, denyIPAddressList } );
+
+        request( app )
+            .get( requestURL )
+            .set( 'Referer', 'https://google.com?print=true' )
+            .expect( 200, testHTMLBody )
+            .end( done );
+    } );
+    it( 'for a private IP address WITH a Referer with allowPrivateIPAddress=true and denyIPAddressList=[`127.0.0.1`]', done => {
+
+        // Change two settings
+        const allowPrivateIPAddress = true;
+        const denyIPAddressList = [localhost];
+
+        app.set( 'ip filtering', { allowPrivateIPAddress, allowMetaIPAddress, allowIPAddressList, denyIPAddressList } );
+
+        request( app )
+            .get( requestURL )
+            .set( 'Referer', 'https://google.com?print=true' )
+            .expect( 500, /DNS lookup .* is not allowed. Because It is defined in denyIPAddressList./ )
+            .end( done );
+    } );
+
+    // Tests WITHOUT Referers
+
+    // Tests with allowPrivateIPAddress FALSE
+    it( 'for a private IP address WITHOUT a Referer with allowPrivateIPAddress=false', ( done ) => {
+
+        // Don't change any default IP filtering setting
+        app.set( 'ip filtering', { allowPrivateIPAddress, allowMetaIPAddress, allowIPAddressList, denyIPAddressList } );
+
+        request( app )
+            .get( requestURL )
+            .expect( 500, /DNS lookup .* is not allowed. Because, It is private IP address./ )
+            .end( done );
+    } );
+    it( 'for a private IP address WITHOUT a Referer with allowPrivateIPAddress=false and allowMetaIPAddress=true', done => {
+ 
+        // Only change one setting
+        const allowMetaIPAddress = true;
+
+        app.set( 'ip filtering', { allowPrivateIPAddress, allowMetaIPAddress, allowIPAddressList, denyIPAddressList } );
+
+        request( app )
+            .get( requestMetaURL  )
+            .expect( 500, /Socket is closed/ )
+            .end( done );
+    } );
+    it( 'for a private IP address WITHOUT a Referer with allowPrivateIPAddress=false but allowIPAddressList=[`127.0.0.1`]', done => {
+ 
+        // Only change one setting
+        const allowIPAddressList = [localhost];
+
+        app.set( 'ip filtering', { allowPrivateIPAddress, allowMetaIPAddress, allowIPAddressList, denyIPAddressList } );
+
+        request( app )
+            .get( requestURL )
+            .expect( 200, testHTMLBody )
+            .end( done );
+    } );
+    it( 'for a private IP address WITHOUT a Referer with allowPrivateIPAddress=false and denyIPAddressList=[`127.0.0.1`]', done => {
+
+        // Only change one setting 
+        const denyIPAddressList = [localhost];
+
+        app.set( 'ip filtering', { allowPrivateIPAddress, allowMetaIPAddress, allowIPAddressList, denyIPAddressList } );
+
+        request( app )
+            .get( requestURL )
+            .expect( 500, /DNS lookup .* is not allowed. Because, It is private IP address/ )
+            .end( done );
+    } );
+
+    // Tests with allowPrivateIPAddress TRUE
+    it( 'for a private IP address WITHOUT a Referer with allowPrivateIPAddress=true', done => {
+
+        // Only change one setting
+        const allowPrivateIPAddress = true;
+
+        app.set( 'ip filtering', { allowPrivateIPAddress, allowMetaIPAddress, allowIPAddressList, denyIPAddressList } );
+
+        request( app )
+            .get( requestURL )
+            .expect( 200, testHTMLBody )
+            .end( done );
+    } );
+    it( 'for a private IP address WITHOUT a Referer with allowPrivateIPAddress=true and allowMetaIPAddress=true', done => {
+
+        // Change two settings
+        const allowPrivateIPAddress = true;
+        const allowMetaIPAddress = true;
+
+        app.set( 'ip filtering', { allowPrivateIPAddress, allowMetaIPAddress, allowIPAddressList, denyIPAddressList } );
+
+        request( app )
+            .get( requestMetaURL )
+            .expect( 200, testHTMLBody )
+            .end( done );
+    } );
+    it( 'for a private IP address WITHOUT a Referer with allowPrivateIPAddress=true and allowIPAddressList=[`127.0.0.1`]', done => {
+
+        // Change two settings
+        const allowPrivateIPAddress = true;
+        const allowIPAddressList = [localhost];
+
+        app.set( 'ip filtering', { allowPrivateIPAddress, allowMetaIPAddress, allowIPAddressList, denyIPAddressList } );
+
+        request( app )
+            .get( requestURL )
+            .expect( 200, testHTMLBody )
+            .end( done );
+    } );
+    it( 'for a private IP address WITHOUT a Referer with allowPrivateIPAddress=true and denyIPAddressList=[`127.0.0.1`]', done => {
+
+        // Change two settings
+        const allowPrivateIPAddress = true;
+        const denyIPAddressList = [localhost];
+
+        app.set( 'ip filtering', { allowPrivateIPAddress, allowMetaIPAddress, allowIPAddressList, denyIPAddressList } );
+
+        request( app )
+            .get( requestURL )
+            .expect( 500, /DNS lookup .* is not allowed. Because It is defined in denyIPAddressList./ )
+            .end( done );
+    } );
+
+} );
